feat:improve csv output
Multiple data columns returned from the Amazon gift registry use a comma
as an internal separator. This commit wraps each column in quotes so
that the internal comma separators do not break text-to-column in Excel.

Additionally, because the double quote character, ", is used to wrap
each output column, all double quote characters within a column are
escaped with an additional double quote character to better adhere to
csv standards.

The end result should be a csv file that opens seemlessly in Excel
with each row having the same output columns.