API fixes, Ease of Use and README updates (#34)
* fix:add newly required 'token' POST request parameter

* fix:update address element selector

* fix:update API response gift date parameter after API change

* feat:handle no address provided by gifter

* feat:add registry constants block for ease of use

Added javascript variables near beginning of script with a note to
update them before running the script to help with ease of use.

Each individual registry will have different parameters and the user
will need to update these variables so that the script will run on their
registry.

* feat:improve csv output

Multiple data columns returned from the Amazon gift registry use a comma
as an internal separator. This commit wraps each column in quotes so
that the internal comma separators do not break text-to-column in Excel.

Additionally, because the double quote character, ", is used to wrap
each output column, all double quote characters within a column are
escaped with an additional double quote character to better adhere to
csv standards.

The end result should be a csv file that opens seemlessly in Excel
with each row having the same output columns.

* feat:parameterize output column separator

Adds a const at the beginning of the script to allow for only one change
being required to change the output column delimiter. E.g. if someone
wants to change the output delimiter to a pipe, |, they can do so with
only one change instead of two.

* Update README.md

* Adds an explanation of how to find the relevant POST parameters in the network tab.
* Adds a screenshot of where the "Export gift details" button is with the current registry UI